refactor: rename steps.steps to steps.types
