+++ app_test_suite/steps/pytest/pytest.py
@@ -15,7 +15,7 @@ from app_test_suite.steps.base_test_runner import (
     BaseTestRunner,
     context_key_chart_yaml,
 )
-from app_test_suite.steps.steps import STEP_TEST_SMOKE, STEP_TEST_FUNCTIONAL
+from app_test_suite.steps.types import STEP_TEST_SMOKE, STEP_TEST_FUNCTIONAL
 from step_exec_lib.errors import ValidationError
 from step_exec_lib.types import Context, StepType
 from step_exec_lib.utils.config import get_config_value_by_cmd_line_option
